Reuse a helper function
